// redux/tag-types.ts
export enum tagTypes {
  auth = "auth",
  logout = "logout",
  category = "category",
  brands = "brands",
  taxRule = "taxRule",
  shipping = "shipping",
  bundleDiscount = "bundleDiscount",
  productCollection = "productCollection",
  voucher = "voucher",
  variants = "variants",
  variantItem = "variantItem",
  product = "product",
  vendor = "vendor",
  carts = "carts",
  user = "user",
  abandonedCart = "abandoned-cart",
  address = "address",
  image = "image",
  inventory = "inventory",
  banner = "banner",
  coupon = "coupon",
  shippingRules = "shippingRules",
  wishlist = "wishlist",
  bannerAd = "bannerAd",
  pageLink = "pageLink",
  career = "career",
  blog = "blog",
  application = "application",
  customDesign = "customDesign",
  feature = "feature",
}

export const tagTypesList = [
  tagTypes.auth,
  tagTypes.logout,
  tagTypes.carts,
  tagTypes.user,
  tagTypes.category,
  tagTypes.brands,
  tagTypes.taxRule,
  tagTypes.shipping,
  tagTypes.bundleDiscount,
  tagTypes.productCollection,
  tagTypes.voucher,
  tagTypes.variants,
  tagTypes.variantItem,
  tagTypes.product,
  tagTypes.vendor,
  tagTypes.abandonedCart,
  tagTypes.address,
  tagTypes.image,
  tagTypes.inventory,
  tagTypes.banner,
  tagTypes.coupon,
  tagTypes.shippingRules,
  tagTypes.bannerAd,
  tagTypes.pageLink,
  tagTypes.career,
  tagTypes.blog,
  tagTypes.application,
  tagTypes.customDesign,
  tagTypes.wishlist,
  tagTypes.feature,
];